% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_color}
\alias{chk_color}
\alias{chk_colour}
\title{Check Color String}
\usage{
chk_color(x, x_name = NULL)

chk_colour(x, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
\code{NULL}, invisibly. Called for the side effect of throwing an error
if the condition is not met.
}
\description{
Checks that x is a string (non-missing character vector of length 1)
that specifies a color.

\code{chk_color}
checks if a color string.

\code{chk_colour}
checks if a color string.
}
\section{Functions}{
\itemize{
\item \code{chk_color}: Check Color String Object

\item \code{chk_colour}: Check Color String Object
}}

\examples{

# chk_color
chk_color("blue")
try(chk_color("glue"))

# chk_colour
chk_colour("blue")
try(chk_colour("glue"))
}
\seealso{
\code{\link[=vld_color]{vld_color()}}
}
