% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinter.R
\name{lighten}
\alias{lighten}
\title{Lighten colour.}
\usage{
lighten(x, amount)
}
\arguments{
\item{x}{A vector of strings of colours in any format accepted by grDevices::col2rgb().}

\item{amount}{A number from 0 to 1.}
}
\value{
A vector of modified colours.
}
\description{
Lighten colour.
}
\examples{
lighten(tinter("blue"), 0.2)
}
