% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vld.R
\name{vld_color}
\alias{vld_color}
\alias{vld_colour}
\title{Validate Color String}
\usage{
vld_color(x)

vld_colour(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A flag indicating whether the object was validated.
}
\description{
Validates whether x is a string (non-missing character vector of length 1)
that specifies a color.
}
\section{Functions}{
\itemize{
\item \code{vld_color}: Validate Color String

\item \code{vld_colour}: Validate Colour String
}}

\examples{

# vld_color
vld_color("blue")
vld_color("glue")

# vld_color
vld_colour("blue")
vld_colour("glue")
}
\seealso{
\code{\link[=chk_color]{chk_color()}}
}
